moved the kernel listener compiler pass to HttpKernel to make it reusable (refs #6643)
